requirements.txt
dragonfly2
dfly-breathe
six
future
wxpython
pywin32
setuptools